PIM-5218: Move BulkVersionSaver
